backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pyyaml==6.0.2
markdown==3.7
python-frontmatter==1.1.0
python-multipart==0.0.9
aiofiles==24.1.0
